refactor: split the two oversized modules along their internal seams
commands/extend.py (1714 lines) and report/html.py (1349) between them held
a third of the package, and neither was one thing.

report/html.py becomes four modules, each with a single job:
  css.py     palette + stylesheet   (the two are one decision)
  charts.py  inline SVG charts
  panels.py  per-read alignment panels
  html.py    document assembly

commands/extend.py keeps the click command and orchestration; the rest moves:
  report/extend_report.py  the Markdown statistics report and the constants
                           bounding what the HTML view retains
  commands/_options.py     argument parsing and validation

count_terminal_motifs deliberately stays in the command module rather than
moving to core or io: it raises click.ClickException, and no core/io/report
module imports click. Keeping that coupling in the command layer is worth more
than the ~90 lines it would shed.

Largest module is now 840 lines, down from 1714.

Two formatting helpers were duplicated between the text and HTML reports.
html._fmt was identical to text.fmt_int and is now that function. html._delta
differed only in rendering zero as a typographic en dash, so fmt_delta gains a
`zero` parameter and text.fmt_delta_html names the HTML convention once rather
than repeating the literal at four call sites.

The compat shim for teloclip.html_report now fronts several modules, so
test_compat_shims.py takes a tuple of source modules per shim. Names that
were private before the split (motif_pattern) are re-exported since they are
public now.
